# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Followup API",
        "version": "1.0.0"
    },
    "basePath": "/api",
    "paths": {
        "/post/{post}/comments": {
            "get": {
                "tags": [
                    "comments"
                ],
                "summary": "post comments",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "objects"
                    }
                }
            },
            "post": {
                "tags": [
                    "comments"
                ],
                "summary": "add new comment",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    },
                    {
                        "name": "user_id",
                        "in": "formData",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    },
                    {
                        "name": "body",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "objects"
                    }
                }
            }
        },
        "comment/{comment}/update": {
            "post": {
                "tags": [
                    "comments"
                ],
                "summary": "update comment",
                "parameters": [
                    {
                        "name": "comment",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    },
                    {
                        "name": "user_id",
                        "in": "formData",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    },
                    {
                        "name": "body",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "objects"
                    }
                }
            }
        },
        "comment/{comment}/delete": {
            "post": {
                "tags": [
                    "comments"
                ],
                "summary": "delete comment",
                "parameters": [
                    {
                        "name": "comment",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    },
                    {
                        "name": "user_id",
                        "in": "formData",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "objects"
                    }
                }
            }
        },
...
        "/post/{post}/likes": {
            "get": {
                "tags": [
                    "likes"
                ],
                "summary": "post likes",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "objects"
                    }
                }
            },
            "post": {
                "tags": [
                    "likes"
                ],
                "summary": "add like",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    },
                    {
                        "name": "user_id",
                        "in": "formData",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "objects"
                    }
                }
            }
        },
        "/like/{like}/delete": {
            "post": {
                "tags": [
                    "likes"
                ],
                "summary": "delete like",
                "parameters": [
                    {
                        "name": "like",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    },
                    {
                        "name": "user_id",
                        "in": "formData",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    },
                    {
                        "name": "post_id",
                        "in": "formData",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "objects"
                    }
                }
            }
        },
...
    },
...
}